Moving some unit test functionality into a more global location
